server/pyproject.toml
[tool.poetry]
name = "universal-widgets"
version = "0.1.0"
description = ""
authors = ["Mahfuz Khan <mahfuzkhanshovon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.15"
requests = "^2.25.1"
Flask = "^2.0.1"
python-dotenv = "^0.17.1"
Flask-SQLAlchemy = "^3.0.2"
PyMySQL = "^1.0.2"
SQLAlchemy-JSONField = "^1.0.0"
pytz = "^2021.3"
rq = "^1.10.1"
gunicorn = "^20.1.0"
humanize = "^3.14.0"
PyJWT = "^2.3.0"
python-dateutil = "^2.8.2"
gevent = "^21.12.0"
phonenumbers = "^8.12.48"
flask_profiler = "^1.8.1"
stripe = "^4.0.2"
twilio = "^7.12.0"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^4.0.4"
python-jose = "^3.3.0"
openai = "^1.13.3"
eventlet = "^0.33.3"
Flask-SocketIO = "^5.3.3"
redis = "^4.5.4"
kombu = "^5.2.4"
Flask-Session = "^0.4.1"
gevent-websocket = "^0.10.1"
paramiko = "^3.1.0"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
plotly = "^5.14.1"
scikit-learn = "^1.2.2"
boto3 = "^1.26.132"
tiktoken = "^0.4.0"
SQLAlchemy-serializer = "^1.4.1"
flask-validate-json = "^0.0.1"
Flask-JWT-Extended = "^4.6.0"
WTForms = "^3.1.2"
PyPDF2 = "^3.0.1"
google-auth = "^2.35.0"
passlib = "^1.7.4"
flask-admin = "^1.6.1"
flask-basicauth = "^0.2.0"
faker = "^37.4.0"
firebase-admin = "^7.1.0"
google-cloud-speech = "^2.33.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
